+++ projects/u7agent/src/features/memory/short-term-memory.ts
